Rename store.rs -> engine.rs
