+++ packages/vm/src/wasm_backend/mod.rs
@@ -1,8 +1,8 @@
 mod compile;
+mod engine;
 mod gatekeeper;
 mod limiting_tunables;
-mod store;
 
 pub use compile::compile;
+pub use engine::{make_compiling_engine, make_runtime_engine};
 pub use limiting_tunables::LimitingTunables;
-pub use store::{make_compiling_engine, make_runtime_engine};
